Update dirichlet_prior_demo.py
I made a few changes for readability. I also adapted the code so the user has to specify the mode. I removed the montecarlo routine, because we can evaluate the cdf directly. 

This will fails if the mass is not very tight (at least for some values of the mode). Still not sure if this is just a numerical issue to improve or a fundamental limitations. I will test it later